Bug 2022734 [wpt PR 58435] - wpt: Use the default Servo user agent for WPT tests, a=testonly
Automatic update from web-platform-tests
wpt: Use the default Servo user agent for WPT tests

Some WPT tests [^1] rely on the user agent reflecting that of a real
browser or platform. This change makes it so that the user agent that
Servo uses when running the WPT is the default one, which includes the
platform.

[^1]: https://github.com/servo/servo/blob/aae5c4cd973a7c2179f9f5da29d287deff124f00/tests/wpt/tests/resources/accesskey.js#L16


--

wpt-commits: 5217b2ef4fae7c859f3e50c44100cba1897619f1
wpt-pr: 58435

UltraBlame original commit: 22b1cb949aa0441136a1d2cb637cbb4f140a01e2

+++ testing/web-platform/tests/tools/wptrunner/wptrunner/browsers/servo.py
@@ -96,7 +96,6 @@ class ServoBrowser(WebDriverBrowser):
 
         args = [
             "--hard-fail",
-            "-u", "Servo/wptrunner",
             
             
             "--ignore-certificate-errors",
@@ -140,6 +139,7 @@ class ServoBrowser(WebDriverBrowser):
         except requests.exceptions.Timeout:
             
             
+            
             self.logger.debug("Servo webdriver status request timed out.")
             return True
         except Exception as exception:
@@ -149,29 +149,34 @@ class ServoBrowser(WebDriverBrowser):
         return True
 
     def stop(self, force=False):
-        retry_cnt = 0
-        while self.is_alive():
+        for _ in range(self.shutdown_retry_attempts):
+            if not self.is_alive():
+                return
+
             self.logger.info("Trying to shut down gracefully by extension command")
             try:
+                
                 requests.delete(
                     f"http://{self.host}:{self.port}/session/dummy-session-id/servo/shutdown",
                     timeout=3
                 )
             except requests.exceptions.ConnectionError:
                 self.logger.debug("Browser already shut down (connection refused)")
-                break
-            except requests.exceptions.RequestException as exeception:
-                self.logger.debug(f"Request exception: {exeception}")
-                break
-            except requests.exceptions.Timeout:
-                self.logger.debug("Request timed out")
+                return
+            except requests.exceptions.RequestException as exception:
+                self.logger.debug(
+                    f"Request exception: {exception}. "
+                    "This normally means webdriver server is waiting synchronously "
+                    "for response of the previous command.\n"
+                    "This always follows 'Servo webdriver status request timed out.'"
+                )
+                self.logger.warning("Hanged server. Killing instead.")
                 break
 
-            retry_cnt += 1
-            if retry_cnt >= self.shutdown_retry_attempts:
-                self.logger.warning("Max retry exceeded to normally shut down. Killing instead.")
-                break
             time.sleep(1)
+        else:
+            self.logger.warning("Max retry exceeded to normally shut down. Killing instead.")
+
         super().stop(force)
 
     def find_wpt_prefs(self, logger):
